VirtualSeco/seco/twitterSkApi/requirements.txt
apiview==1.3.25
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorators==2.0.7
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
joblib==1.0.0
MarkupPy==1.14
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
pycparser==2.20
PyMySQL==1.0.2
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
serializers==0.2.4
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
textblob==0.15.3
tqdm==4.56.0
tweepy==3.10.0
urllib3==1.26.3
views==0.3
xlrd==2.0.1
xlwt==1.3.0
